Land ESM-based config support without requiring opt-in
