+++ lib/eslint-plugin-helper-worker.js
@@ -1,0 +1,83 @@
+'use strict';
+const v8 = require('v8');
+const {parentPort, workerData} = require('worker_threads');
+
+const {normalizeGlobs} = require('./globs');
+const normalizeExtensions = require('./extensions');
+const {loadConfig} = require('./load-config');
+const providerManager = require('./provider-manager');
+
+const MAX_DATA_LENGTH_EXCLUSIVE = 100 * 1024; // Allocate 100 KiB to exchange globs.
+
+const configCache = new Map();
+
+const collectProviders = ({conf, projectDir}) => {
+	const providers = [];
+	if (Reflect.has(conf, 'babel')) {
+		const {level, main} = providerManager.babel(projectDir);
+		providers.push({
+			level,
+			main: main({config: conf.babel}),
+			type: 'babel'
+		});
+	}
+
+	if (Reflect.has(conf, 'typescript')) {
+		const {level, main} = providerManager.typescript(projectDir);
+		providers.push({
+			level,
+			main: main({config: conf.typescript}),
+			type: 'typescript'
+		});
+	}
+
+	return providers;
+};
+
+const buildGlobs = ({conf, providers, projectDir, overrideExtensions, overrideFiles}) => {
+	const extensions = overrideExtensions ?
+		normalizeExtensions(overrideExtensions) :
+		normalizeExtensions(conf.extensions, providers);
+
+	return {
+		cwd: projectDir,
+		...normalizeGlobs({
+			extensions,
+			files: overrideFiles ? overrideFiles : conf.files,
+			providers
+		})
+	};
+};
+
+const resolveGlobs = async (projectDir, overrideExtensions, overrideFiles) => {
+	if (!configCache.has(projectDir)) {
+		configCache.set(projectDir, loadConfig({resolveFrom: projectDir}).then(conf => { // eslint-disable-line promise/prefer-await-to-then
+			const providers = collectProviders({conf, projectDir});
+			return {conf, providers};
+		}));
+	}
+
+	const {conf, providers} = await configCache.get(projectDir);
+	return buildGlobs({conf, providers, projectDir, overrideExtensions, overrideFiles});
+};
+
+const data = new Uint8Array(workerData.dataBuffer);
+const sync = new Int32Array(workerData.syncBuffer);
+
+const handleMessage = async ({projectDir, overrideExtensions, overrideFiles}) => {
+	let encoded;
+	try {
+		const globs = await resolveGlobs(projectDir, overrideExtensions, overrideFiles);
+		encoded = v8.serialize(globs);
+	} catch (error) {
+		encoded = v8.serialize(error);
+	}
+
+	const byteLength = encoded.length < MAX_DATA_LENGTH_EXCLUSIVE ? encoded.copy(data) : MAX_DATA_LENGTH_EXCLUSIVE;
+	Atomics.store(sync, 0, byteLength);
+	Atomics.notify(sync, 0);
+};
+
+parentPort.on('message', handleMessage);
+handleMessage(workerData.firstMessage);
+delete workerData.firstMessage;
