+++ test-tap/fixture/eslint-plugin-helper/for-overriding/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
