+++ test-tap/fixture/eslint-plugin-helper/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
