+++ test-tap/fixture/load-config/contains-ava-property/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
