+++ test-tap/fixture/load-config/factory-no-plain-return/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
