--- test-tap/fixture/load-config/factory-no-promise-return/ava.config.js
@@ -1,5 +1,0 @@
-export default () => new Promise(resolve => {
-	resolve({
-		files: 'this-should-not-work'
-	});
-});
