--- test-tap/fixture/load-config/factory-no-promise-return/package.json
@@ -1,1 +1,0 @@
-{}
