+++ test-tap/fixture/load-config/file-yes-cjs-yes/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
