--- test-tap/fixture/load-config/mjs/ava.config.mjs
@@ -1,5 +1,0 @@
-export default ({projectDir}) => {
-	return {
-		files: projectDir
-	};
-};
