--- test-tap/fixture/load-config/mjs/package.json
@@ -1,1 +1,0 @@
-{}
