+++ test-tap/fixture/load-config/no-default-export/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
