+++ test-tap/fixture/load-config/no-plain-config/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
