--- test-tap/fixture/load-config/no-promise-config/ava.config.js
@@ -1,1 +1,0 @@
-export default Promise.resolve('should not work!');
