--- test-tap/fixture/load-config/no-promise-config/package.json
@@ -1,1 +1,0 @@
-{}
