+++ test-tap/fixture/load-config/non-object-experiments/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
