+++ test-tap/fixture/load-config/package-no-file-yes/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
