+++ test-tap/fixture/load-config/package-yes-explicit-yes/package.json
@@ -4,5 +4,6 @@
 			"abc",
 			"!xyz"
 		]
-	}
+	},
+	"type": "module"
 }
