+++ test-tap/fixture/load-config/package-yes-file-yes/package.json
@@ -4,5 +4,6 @@
 			"abc",
 			"!xyz"
 		]
-	}
+	},
+	"type": "module"
 }
