+++ test-tap/fixture/load-config/require/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
