+++ test-tap/fixture/load-config/throws/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
