+++ test-tap/fixture/load-config/unsupported-experiments/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
