+++ test/config/fixtures/factory-promise-return/ava.config.js
@@ -1,8 +1,5 @@
 module.exports = async () => {
 	return {
-		nonSemVerExperiments: {
-			nextGenConfig: true
-		},
 		failFast: true
 	};
 };
