+++ test/config/fixtures/file-yes-cjs-yes-mjs-yes/ava.config.mjs
@@ -1,6 +1,3 @@
 export default {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	files: 'package-yes-files-yes-test-value'
 };
