+++ test/config/fixtures/file-yes-cjs-yes-mjs-yes/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
