+++ test/config/fixtures/js-as-cjs/ava.config.js
@@ -1,6 +1,3 @@
 module.exports = {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 };
