+++ test/config/fixtures/js-as-esm/ava.config.js
@@ -1,6 +1,3 @@
 export default {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 };
