+++ test/config/fixtures/js-as-esm/error.js
@@ -1,7 +1,4 @@
 export default {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 };
