+++ test/config/fixtures/js-as-esm/no-default-export.js
@@ -1,6 +1,3 @@
 export const config = {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 };
