+++ test/config/fixtures/mjs-with-tests/ava.config.mjs
@@ -1,6 +1,3 @@
 export default {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	files: ['dir-a/*.js']
 };
