+++ test/config/fixtures/mjs/ava.config.mjs
@@ -1,6 +1,3 @@
 export default {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 };
