+++ test/config/fixtures/mjs/error.mjs
@@ -1,7 +1,4 @@
 export default {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 };
