+++ test/config/fixtures/mjs/no-default-export.mjs
@@ -1,6 +1,3 @@
 export const config = {
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 };
