+++ test/config/fixtures/promise-config/ava.config.js
@@ -1,6 +1,3 @@
 module.exports = Promise.resolve({
-	nonSemVerExperiments: {
-		nextGenConfig: true
-	},
 	failFast: true
 });
