+++ test/config/snapshots/loader.js.md
@@ -28,12 +28,6 @@ Generated by [AVA](https://avajs.dev).
 
     'Error loading ava.config.js: require is not defined'
 
-## throws an error if a config factory returns a promise
-
-> error message
-
-    'ava.config.js exported a promise or an asynchronous factory function. You must enable the ’asyncConfigurationLoading’ experiment for this to work.'
-
 ## throws an error if a config exports a promise
 
 > error message
@@ -56,7 +50,7 @@ Generated by [AVA](https://avajs.dev).
 
 > error message
 
-    'ava.config.js must have a default export, using ES module syntax'
+    'Error loading ava.config.js: ava.config.js must have a default export'
 
 ## throws an error if a config file contains `ava` property
 
@@ -81,9 +75,3 @@ Generated by [AVA](https://avajs.dev).
 > error message
 
     'Conflicting configuration in ava.config.js and ava.config.cjs'
-
-## refuses to load .mjs config
-
-> error message
-
-    'AVA cannot yet load ava.config.mjs files'
