+++ test/config/snapshots/next-gen.js.md
@@ -8,13 +8,13 @@ Generated by [AVA](https://avajs.dev).
 
 > error message
 
-    'Error loading error.mjs: 🙈'
+    'Error loading error.mjs'
 
 ## fails when .mjs config does not have a default export
 
 > error message
 
-    'no-default-export.mjs must have a default export'
+    'Error loading no-default-export.mjs'
 
 ## handles errors when loading .js config as ESM
 
@@ -26,7 +26,7 @@ Generated by [AVA](https://avajs.dev).
 
 > error message
 
-    'no-default-export.js must have a default export'
+    'Error loading no-default-export.js: no-default-export.js must have a default export'
 
 ## throws an error if .js, .cjs and .mjs configs are present
