+++ test/configurable-module-format/fixtures/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
