+++ test/environment-variables/fixtures/environment-variables/environment-variables.js
@@ -1,4 +1,4 @@
-const test = require('ava');
+import test from 'ava';
 
 test('works', t => {
 	t.is(process.env.MY_ENVIRONMENT_VARIABLE, 'some value');
