+++ test/environment-variables/fixtures/environment-variables/package.json
@@ -1,1 +1,3 @@
-{}
+{
+	"type": "module"
+}
